Move all discord related listener into me.mastercapexd.auth.discord.listener
