+++ core/src/main/java/me/mastercapexd/auth/discord/command/DiscordCommandRegistry.java
@@ -3,23 +3,19 @@ package me.mastercapexd.auth.discord.command;
 import java.util.Collections;
 
 import com.bivashy.auth.api.AuthPlugin;
-import com.bivashy.auth.api.config.link.command.LinkCommandPathSettings;
 
-import me.mastercapexd.auth.config.discord.DiscordCommandSettings;
 import me.mastercapexd.auth.discord.command.annotation.RenameTo;
-import me.mastercapexd.auth.discord.command.listener.JDACommandListener;
+import me.mastercapexd.auth.discord.listener.JDACommandListener;
 import me.mastercapexd.auth.hooks.DiscordHook;
 import me.mastercapexd.auth.link.LinkCommandActorWrapper;
 import me.mastercapexd.auth.link.discord.DiscordCommandActorWrapper;
 import me.mastercapexd.auth.link.discord.DiscordLinkType;
 import me.mastercapexd.auth.messenger.commands.MessengerCommandRegistry;
-import me.mastercapexd.auth.messenger.commands.annotation.CommandKey;
 import me.mastercapexd.auth.shared.commands.DiscordLinkCommand;
 import me.mastercapexd.auth.shared.commands.MessengerLinkCommandTemplate;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import revxrsal.commands.annotation.dynamic.Annotations;
 import revxrsal.commands.command.CommandActor;
-import revxrsal.commands.command.ExecutableCommand;
 import revxrsal.commands.jda.JDAActor;
 import revxrsal.commands.jda.JDACommandHandler;
 import revxrsal.commands.jda.annotation.OptionData;
